feat(admin): 2261 - Centres - Mise à jour selecteur de cohort (split)
